// src/organisms/ce-mapping-experts/types.ts
export type RowsData = {
    id: number;
    expert_name?: string;
    status: string;
    current_company_name: string;
    current_designation: string;
    relevant_company_name: string;
    relevant_designation: string;
    expert_base_location: string;
    action: boolean;
    added_on: string | null;
    last_updated_on: string | null;
}